Tercera versión: Se agrego al servicio de delete una verificación para ver si el id que se va a eliminar existe en la lista con la exception ResourceNotFoundException, se agrego metodo para ver la venta total de los boletos, y se agregaron test Unitarios a cada ServiceImpl todos los tests con el 100%

+++ src/main/java/com/metahorce/cinemagic/services/CalificacionServiceImpl.java
@@ -63,7 +63,7 @@ public class CalificacionServiceImpl implements CalificacionService {
     public Calificacion updateCalificacion(Integer id, Calificacion calificacion){
         Calificacion calificacionUpdate = calificacionRepository.findById(id).orElse(null);
         if (calificacionUpdate == null){
-            throw new ResourceNotFoundException("No se encontro la calificacón con el id: "+ id);
+            throw new ResourceNotFoundException("No se encontro la calificación con el id: "+ id);
         }
 
         Integer idUsuario = calificacion.getUsuario().getId();
@@ -87,6 +87,9 @@ public class CalificacionServiceImpl implements CalificacionService {
 
     @Override
     public void deleteCalificacion(Integer id){
+        if (!calificacionRepository.existsById(id)) {
+            throw new ResourceNotFoundException("No se encontro la calificación con el id: " + id);
+        }
         calificacionRepository.deleteById(id);
     }
